Put createLabel and createLabelClass in common declaration file
