+++ cypress/integration/polygon-drawing.online.ts
@@ -1,43 +1,10 @@
-import { gql } from "@apollo/client";
-
 import { distantDatabaseClient as client } from "../../typescript/web/src/connectors/apollo-client/client";
-import { createDataset, createImage } from "./graphql-definitions.common";
+import {
+  createDataset,
+  createImage,
+  createLabelClass,
+} from "./graphql-definitions.common";
 import { declarePolygonDrawingTests } from "./polygon-drawing.common";
-
-const createLabelClass = async (
-  name: String,
-  color = "#ffffff",
-  datasetId: string
-) => {
-  const {
-    data: {
-      createLabelClass: { id },
-    },
-  } = await client.mutate({
-    mutation: gql`
-      mutation createLabelClass(
-        $name: String!
-        $color: String!
-        $datasetId: ID!
-      ) {
-        createLabelClass(
-          data: { name: $name, color: $color, datasetId: $datasetId }
-        ) {
-          id
-          name
-          color
-        }
-      }
-    `,
-    variables: {
-      name,
-      color,
-      datasetId,
-    },
-  });
-
-  return id;
-};
 
 describe("Polygon drawing (online)", () => {
   let datasetId: string;
@@ -64,7 +31,12 @@ describe("Polygon drawing (online)", () => {
       });
       imageId = id;
 
-      await createLabelClass("Rocket", "#F87171", datasetId);
+      await createLabelClass({
+        name: "Rocket",
+        color: "#F87171",
+        datasetId,
+        client,
+      });
     });
   });
